mv s3_stub to conftest
