+++ lambdas/s3hash/tests/test_compute_checksum.py
@@ -1,4 +1,3 @@
-import asyncio
 import base64
 import io
 
@@ -32,14 +31,6 @@ def make_body(contents: bytes) -> StreamingBody:
     )
 
 
-AWS_CREDENTIALS = s3hash.AWSCredentials.parse_obj(
-    {
-        "key": "test-key",
-        "secret": "test-secret",
-        "token": "test-token",
-    }
-)
-
 LOC = s3hash.S3ObjectSource(
     bucket="test-bucket",
     key="test-key",
@@ -56,24 +47,6 @@ EXPECTED_MPU_PARAMS = {
     **s3hash.MPU_DST.boto_args,
     "ChecksumAlgorithm": "SHA256",
 }
-
-
-# pytest's async fixtures don't propagate contextvars, so we have to set them manually in a sync fixture
-@pytest.fixture
-def s3_stub():
-    async def _get_s3():
-        async with s3hash.aio_context(AWS_CREDENTIALS):
-            return s3hash.S3.get()
-
-    s3 = asyncio.run(_get_s3())
-    stubber = Stubber(s3)
-    stubber.activate()
-    s3_token = s3hash.S3.set(s3)
-    try:
-        yield stubber
-        stubber.assert_no_pending_responses()
-    finally:
-        s3hash.S3.reset(s3_token)
 
 
 async def test_compliant(s3_stub: Stubber):
